+++ tests/Executor/ValidationWithCacheTest.php
@@ -8,6 +8,7 @@ use GraphQL\GraphQL;
 use GraphQL\Language\AST\DocumentNode;
 use GraphQL\Tests\Executor\TestClasses\Cat;
 use GraphQL\Tests\Executor\TestClasses\Dog;
+use GraphQL\Tests\Executor\TestClasses\SpyValidationCacheAdapter;
 use GraphQL\Tests\PsrValidationCacheAdapter;
 use GraphQL\Type\Definition\InterfaceType;
 use GraphQL\Type\Definition\ObjectType;
@@ -16,25 +17,6 @@ use GraphQL\Type\Schema;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Cache\Adapter\ArrayAdapter;
 use Symfony\Component\Cache\Psr16Cache;
-
-final class SpyValidationCacheAdapter extends PsrValidationCacheAdapter
-{
-    public int $isValidatedCalls = 0;
-    public int $markValidatedCalls = 0;
-
-    public function isValidated(Schema $schema, DocumentNode $ast): bool
-    {
-        $this->isValidatedCalls++;
-        return parent::isValidated($schema, $ast);
-    }
-
-    public function markValidated(Schema $schema, DocumentNode $ast): void
-    {
-        $this->markValidatedCalls++;
-        parent::markValidated($schema, $ast);
-    }
-}
-
 
 final class ValidationWithCacheTest extends TestCase
 {
